Working version of local_2_cars, added configuration file for experiment parameters. Still need to work on global function.
