+++ config/config.ini
@@ -1,0 +1,17 @@
+
+[ExperimentSettings]
+experiment_log_directory = exp4
+load_weight_directory = experiments/exp2/weights/12_sixth_right.h5
+save_weight_current_directory = /1_first_left.h5
+
+car1_location = -20, 0
+car2_location = 0, -20
+
+global_experiment = False
+max_episodes = 10
+max_steps = 500
+
+alternate_training = True
+alternate_car = 1
+
+
